feat: implement WiFi config and HomeKit separation for stable pairing
- Separate WiFi configuration mode from HomeKit operation mode
- WiFi config only runs when no configuration exists (AP mode)
- HomeKit mode runs in clean environment without WebServer interference
- Fix HomeKit pairing issues by proper initialization sequence
- Use fixed pairing code "11122333" for stability
- Hardware initialization before HomeKit setup
- Remove WebServer after WiFi connection to avoid conflicts

+++ include/protocol/S21Utils.h
@@ -22,8 +22,8 @@ static inline float s21_decode_target_temp(unsigned char v) {
     return 18.0 + 0.5 * ((signed)v - AC_MIN_TEMP_VALUE);
 }
 
-static inline float s21_encode_target_temp(float temp) {
-    return lroundf((temp - 18.0) * 2) + AC_MIN_TEMP_VALUE;
+static inline uint8_t s21_encode_target_temp(float temp) {
+    return (uint8_t)(lroundf((temp - 18.0) * 2) + AC_MIN_TEMP_VALUE);
 }
 
 static inline int s21_decode_int_sensor(const unsigned char* payload) {
